Use `setup_test_component_platform` helper for binary_sensor entity component tests instead of `hass.components`
